Optimized added parent fanout: New deps NotSubtype
Summary:
This diff and following two diffs aim at fixing the following edge case in the new added parent fanout algo:
```
// a.php
class A {}

// b.php
class B {}

// c.php
function f(A $a): void {
  if ($a is B) {
    // We refine the type of $a to A & B
    expect_nothing($a)
  }
}

function expect_nothing(nothing $_): void {}
```
When changing A:
```
class A extends B {}
```
`f` will not be rechecked by the fanout algo, even though it now has an error.

### What's really going on here?

The code above typechecks (before user change) using facts of a kind we haven't considered when deriving the algorithm: that A is not a subtype of B and B is not a subtype of A. So "is not subtype" kind of facts: not (t1 <: t2).

Obviously, adding a parent can invalidate some of these facts, so whatever typing derivation we did using those facts needs to be redone, and the file (we did the typing derivation for) needs to be included in the fanout.

### Fix

We can use a new kind of dependency edge `NotSubtype` whenever we use a 'is not subtype' fact in a typing derivation. So in the example above, we used not (A <: B) to derive that f typechecks, so we'd record a NotSubtype edge between dependent f and dependency A.

Then when adding a parent to some type X, we'd gather all the descendants of X via the Extends edges of the graph, then follow any NotSubtype edge from any of these descendants and include the end of those edges in the fanout.

### This diff
just introduces the new kind of edge NotSubtype. Similar to the `RequireExtends` edge, we obtain the hash for NotSubtype A by rehashing the hash for `Type A` with `KNotSubtype`. This allows appropriate traversal (in next diff).

Differential Revision: D51031415

fbshipit-source-id: 8b63e5158974ebc42d5de2da95e1935f39d70a20