Add full AnalysisIndex to HHBBC
Summary:
Add the full implementation of AnalysisIndex to HHBBC. This is the
equivalent of Index, but for extern-worker analysis jobs. Not all
functionality is implemented, and not all cases are handled, but the
support is sufficient for the analyze constants pass. Like Index,
AnalysisIndex implements the same IIndex interface, so much of the
generic code works with both. Much of AnalysisIndex's implementation
is based off of Index, but with modifications to handle the possible
lack of data locally.

The only part left to implement is the analysis job itself, which will
tie together the AnalysisIndex and the dependency/scheduler logic.

Reviewed By: jano

Differential Revision: D51387313

fbshipit-source-id: adbdd8688f3c4d025489b9d68c724b2c5cbef57e