Add extern-worker analysis dependency tracking support to HHBBC
Summary:
This begins the adding of support for performing analysis
rounds in an extern-worker job instead of locally. This diff
concentrates on adding the infrastructure to support tracking
dependencies and scheduling work. Most of this infrastructure is not
yet used, but will be in later diffs. It also adds a bare-bones
AnalysisIndex, which will be expanded on in later diffs.

Reviewed By: kavoor

Differential Revision: D51373285

fbshipit-source-id: 7bb49b5cf4b29cd94dcef3115e876b7e5b6afe9c