Elaborate (but don't typecheck) top-level statements in typing_check_job code path
Summary:
Applies elab and validation passes to top-level statements when we are checking the whole program.

Note that the current state of the codebase is such that we have two entrypoints to type-checking, which are used in different places
- typing_toplevel: here we type-check top-level statements
- typing_check_job: we type-check top-level statements neither before nor in this diff

There's a minor user-facing change to some error messages for reserved words in modules, to be addressed separately in T170932344.

Reviewed By: mheiber

Differential Revision: D51590501

fbshipit-source-id: f4325c360e6d3e5e0bc25fea9f968457c93df01e