fix parsing of `default` mode name
Summary: The `default` module name cannot appear in a compound module name.  This was enforced correctly, but the parser was generating odd ASTs which resulted in weird error messages.  This diff updates the parser logic so that a clear error is reported and the AST does not include spurious top-level statements.

Reviewed By: jamesjwu

Differential Revision: D51585375

fbshipit-source-id: 185033445f2e8a28a2a6a197e47bb4da5536b29b