Add full AnalysisIndex to HHBBC (v2)
Summary:
Add the full implementation of AnalysisIndex to HHBBC. This is the
equivalent of Index, but for extern-worker analysis jobs. Not all
functionality is implemented, and not all cases are handled, but the
support is sufficient for the analyze constants pass. Like Index,
AnalysisIndex implements the same IIndex interface, so much of the
generic code works with both. Much of AnalysisIndex's implementation
is based off of Index, but with modifications to handle the possible
lack of data locally.

The only part left to implement is the analysis job itself, which will
tie together the AnalysisIndex and the dependency/scheduler logic.

This diff was responsible for the bugs, as it changed the context
class detection logic in subtle ways for NewObjD.

Reviewed By: kavoor

Differential Revision: D51540432

fbshipit-source-id: b6f7e1c511e9be37c7f16a452ef1a8ecab2c67e4