Add extern-worker analysis dependency tracking support to HHBBC (v2)
Summary:
This begins the adding of support for performing analysis
rounds in an extern-worker job instead of locally. This diff
concentrates on adding the infrastructure to support tracking
dependencies and scheduling work. Most of this infrastructure is not
yet used, but will be in later diffs. It also adds a bare-bones
AnalysisIndex, which will be expanded on in later diffs.

This diff was not responsible for buggs, so unchanged

Reviewed By: kavoor

Differential Revision: D51540370

fbshipit-source-id: c18c62ae9468fc1a861e0ba95b2982e1c8886733