Type UNSAFE_CAST correctly even without everything_sdt
Summary:
Currently, the type of `UNSAFE_CAST` is only correct when pessimised: its type parameters should support dynamic, and its return type should be a like-type. Fix the signature so that it is sound even without the `everything_sdt` flag set.

We need to make use of `~` in `.hhi` files even if `--like-type-hints` is not set. So we change `Naming_validate_like_hint` to support this.

Reviewed By: vassilmladenov

Differential Revision: D51559907

fbshipit-source-id: ab91574b26a55b9a0bf566ddf7e52e2309884517