typing respects cancellation
Summary:
We used to support cancellation by
1. serverMain process would set a bit in shmem
2. whenever a worker read/wrote a decl in shmem, the hh_shared.c could would raise WorkerCancel.Worker_should_exit

That got lost with the move to rust sharded shmem. And also I want cancellation in IDE, which uses localmem backend i.e. doesn't even use sharded shmem. So my idea instead is to place that "raise" check in ocaml code in the few important inner-loops. That's what this diff does.

Which are the important inner loops? I don't know. As a start, I placed the check once per top-level entity and once per member. I also observed that when typechecking EntBusinessUser.php we spend two minutes in `check_members_from_all_parents` so I placed one there. With the checks present in this diff, all my attempted cancellations were respected within a millisecond or so.

(This will also have the property that we can interrupt even if our computation never needs to access the shmem decl store -- as will be typical for didChange or autocomplete typechecking, where all decls are already in local worker memory and all we're doing is typechecking).

Reviewed By: mheiber

Differential Revision: D51543506

fbshipit-source-id: 06f6bff6f592adf4dba8e6667cbe366ac300c42e