refactor Deferred_check
Summary:
This is a no-op refactor in ClientLsp. I recently added a feature `lsp_pull_diagnostics` which makes it so that didOpen/didChange don't immediately trigger a squiggle check; instead they merely stick the URI in a list of "uris which need squiggles recomputed", and we only get around to that after we've done all the more important work like autocomplete. This brought slight improvement to autocomplete times: https://fburl.com/scuba/hh_server_events/4w035o0y
 {F1158245846}

This diff refactors the code slightly.
* Previously `ClientLsp.get_next_event` would return `Deferred_check` indicating that there existed a member in `uris_that_need_check`, and the subsequent invocation of `handle_deferred_check` would assert that `uris_that_need_check` is non-empty and pick+remove an item from it.
* Now, `ClientLsp.get_next_event` actually picks the item and returns `Deferred_check of item`. Therefore, the subsequent invocation of `handle_deferred_check` no longer needs to assert; it merely needs to remove.

This diff is in preparation for further down the stack, where we'll want `handle_deferred_check` to be cancellable. For it to be cancellable, we need to know in advance which URI it's operating upon (hence which subsequent messages might cancel it).

Reviewed By: mheiber

Differential Revision: D51563701

fbshipit-source-id: e7ae28b0a2a77d74d6f19572f6e253649175b4ee