100% rollout lsp_pull_diagnostics, delete flag
Summary:
The lsp_pull_diagnostics JK has been at 50% for two weeks D51258936 and telemetry (in that diff) all shows (1) no problems, (2) a clear win.

The JK gets rolled to 100% here: D51587903

This diff hard-codes the feature to 100% in our codebase. (concomitant cleanup will follow in the next diff in the stack).

Reviewed By: mheiber

Differential Revision: D51432631

fbshipit-source-id: 6048c5a23590a8855d6098515b953f4808fb22b6