Make sure to check modularity for classname_to_class
Summary:
Was reading the implementation for `ClassGetC` for something unrelated and realized I made an error here.

https://www.internalfb.com/code/fbsource/[1698ee8c2346a8d1267aef5d33719c5b0714fe6e]/fbcode/hphp/runtime/vm/bytecode.cpp?lines=2246-2252
https://www.internalfb.com/code/fbsource/[1698ee8c2346]/fbcode/hphp/runtime/vm/bytecode.cpp?lines=980-998
 ---
For the `KindOfString` case of `HH\classname_to_class`, we need to use `Class::resolve` instead of `Class::load`, as `resolve` performs an additional module check.

https://www.internalfb.com/code/fbsource/[6bbe97ded6f8ae1cf1291c866bb9d72f7595adfc]/fbcode/hphp/runtime/vm/class.cpp?lines=5086-5090

 Otherwise, changing
```
$c::foo();
```
to
```
$cl = HH\classname_to_class($c);
$cl::foo();
```
would be a regression in module enforcement.

Reviewed By: oulgen

Differential Revision: D51096864

fbshipit-source-id: 19db2fa39ba9d8f6ce74c1e6a98ff866535783d1