Add JIT implementation for classname_to_class and class_to_classname
Summary: Follow up on review of D51096864, adding jit implementations of class_to_classname and classname_to_class to address concerns about the performance of modularity checking. The former is trivial. For the latter, I add an enum of extra data to the `LdCls` and `LdClsCached` opcodes that governs whether we raise an unrecoverable fatal or an catchable exception when the class does not exist.

Reviewed By: arnabde03

Differential Revision: D51144984

fbshipit-source-id: 4c7237d7756f4d264e2d1f5abbcb2e3e42fbf615