TAST dumping: truncate TAST file names to 255 characters
Summary: With flag `dump_tasts`, we dump each TAST to a different file and create a file name for it using the initial php file path. However, sometimes file names created this way are two long. In this diff, we truncate them to their last 255 characters, which is the file name max length in Linux.

Reviewed By: nt591

Differential Revision: D51660736

fbshipit-source-id: ac5257b101354a9e22b890614111f61ad2db6c95