Get sortText from User Attribute
Summary: In this diff, I take the AutocompleteSortText user attribute I defined and plumb it through all of our decl code! This allows me to fetch it from the class element, then use it as the sort string in clientLsp's hack_to_sort_text function. See the next two diffs in this stack for how it all comes together.

Reviewed By: nt591

Differential Revision: D51318649

fbshipit-source-id: edc9e4375325ee87df99597fd1edff71fcb4a5cf