Move validation for no-top-level-statements from the parser to elaboration
Summary:
## What

User-facing change:
- The error message for top-level statements is now prefixed with  `[Naming]` instead of `[Parsing]`
- In fringe cases of very ill-formed syntax, users won't see the error for top-level statements anymore. For example, in a file consisting only of the `<?hh` line followed by `{`. That's because `hh` and `ClientIdeDaemon` both filter out non-parser errors when there are parse errors  ([source](https://www.internalfb.com/diff/D45174472?permalink=3683905451855273)). I think this is fine–it's unlikely the user was even trying to write a top-level statement in such cases.
- Fix T170543625 : We now show type errors for the rest of a file even if a file contains a top-level statement. This is desirable for Hack notebooks.

## How

Moved the validation for top-level statements from the parser to elab. Implemented in both OCaml and in Rust.

This resolves T170543625 because:
- Hack notebooks user the Hack language server
- Our language server filters out type errors when there are parse errors [since -D45174472](https://www.internalfb.com/diff/D45174472?permalink=3683905451855273). As does `hh`.

## Better engineering

The original design was for the parser to be permissive and for validation to happen during elaboration. This change brings reality closer to that vision for what each pass should do (per enetsee).

## Alternative considered

We could special-case in the filtering logic, but we'd need to do it in two places (language server and `hh_server`) and the change would be unprincipled. The changes in this diff add no special-casing.

## Notes on tests

Some of the nast test snapshots changed to reflect that a parsing error became a naming error. For each of the nast tests with changed snapshots I:
- Inspected the nast with a version of `hh_single_type_check` built off master and `--nast` to verify that there are indeed top-level statements in the nast.
- tested with prod `hh` to verify that the errors did not change, modulo whether they are called `Parsing` or `Naming` errors.

Reviewed By: enetsee

Differential Revision: D51562313

fbshipit-source-id: e0c9df80124381632eafc1d00e7ebacc4b3adadc