Remove extraneous semicolon from thrift/lib/cpp2/protocol/TableBasedSerializerImpl.cpp
Summary:
Extraneous semicolons are a code smell and can mask more serious problems. `-Wextra-semi-stmt` finds them.

This diff removes an extraneous semicolon or adjusts a macro so that a semicolon is required after the macro (making it look like a standard function).

This file is drawn from a heavy-hitting list, so fixing this problem will allow *many* other files to take advantage of the safety `-Wextra-semi-stmt` offers.

This should be a low-risk diff: if it compiles, it works.

Reviewed By: thedavekwon

Differential Revision: D51631143

fbshipit-source-id: d78a6f182c13b6848f4c87d37a5062f218161689