Don't fetch old decls for added and removed symbols
Summary:
This is the more involved diff of the stack.

When computing the fanout of a def, we first determine whether it was added / removed / modified. To know whether it was added or removed, we compare the old and new version of the decl, and if one of the two is absent, we conclude it was added or removed. However, we shouldn't have to obtain the old or new decl in that case, because we already know from the list of names (which we can get from the old and new naming tables) which names have been added or removed. This diff implements that.

Benefits:
* less decling / old decl fetching
* telemetry
  * we're now able to differentiate between a def that was added vs a def for which we're missing the old decl.
  * we can also spot some more pathological cases like when the new decl is missing, which would probably be a bug.

Implementation:
This diff was mostly done by changing the types of things we handle, from mere `path -> FileInfo.names` maps to now maps of `path -> VersionedNames.t`, which implies that other types are changed to one of the new types in Decl_compare. The name diffing logic is confined to those new small modules, as well as some type conversion logic.

Reviewed By: bobrenjc93

Differential Revision: D51554731

fbshipit-source-id: 3f9a6349aa3804fc12fc0820f80d5c12d21c573a