Interpret __DynamicallyCallable on functions as implying __SupportDynamicType
Summary: A function that is marked `__DynamicallyCallable` can be called through a representation of the function that is typed as `dynamic`, or dynamically created from the string name of the function. To be safe, it's necessary for such a function to support dynamic. Rather than require the user to write `__SupportDynamicType` on the function (when `everything_sdt=false`), we can simply assume it in the case that the function is marked `__DynamicallyCallable`.

Reviewed By: mheiber

Differential Revision: D51711364

fbshipit-source-id: b239433bfd0abe5e1964a65fb3a9bde140650564