Housekeeping: rename bridge module compile_ffi -> ffi
Summary:
This is a quality of life change. Naming the bridge module `compile_ffi`
breaks the cxx idiom of calling it `ffi`, and since it's a private module, short & punchy
is better. But the enclosing module is `compiler_ffi`... so it's also better to avoid
a stuttering name with a nearly but not quite the same spelling.

Removed some dead traits while I was in the area.

Reviewed By: nt591

Differential Revision: D51758669

fbshipit-source-id: 09ed60e6152109dfb7fee24fbe4a54b38e5affb9