facts json -> ffi_bridge
Summary:
Right now, the path from a hackc::facts::Facts object to HPHP::Facts::FileFacts
is too long and slow. And worse, hackc::facts::Facts is overly constrained, so it's
too hard to fix it.

1. hackc::facts::Facts -> serde_json::Value
2. serde_json::Value -> json string
3. [optional memcache get/put]
4. json string -> folly::dynamic
5. folly::dynamic -> HPHP::Facts::FileFacts

problem 1: this path is inefficient because there are too many representation conversions, and because json itself is inefficient.

problem 2: The particular json schema used in steps 1 and 5 has a lot of historical
baggage and requires a bunch of custom logic & serde attributes in
hackc::facts::Facts to get the desired json encoding. These customizations
prevent hackc's Facts from working with bincode - it *only* works with json.

problem 3: hackc's Facts contains a number of datatypes, like enums and BTreeMap, that are incompatible with open `cxx` bridge types; cxx cannot generate C++ bindings for them. This, combined with the handcrafted json constriants, make the datatypes hard to work with and hard to change.

observe that the custom json schema, and the requirements for `cxx` compatible datatypes, go hand-in-hand; they're both only for the hackc -> ext_facts code path. It took a lot of hard work to remove other tech-debt tendrils that depended on Facts to get to this point, but now that we're here we have a path to get rid of the handrolled json schema, and move off json entirely.

solution 1: This diff introduces a new set of `cxx` friendly data types that will eventually allow removing most of these conversion steps by allowing C++ to directly access them without converting to folly::dynamic or HPHP::Facts::Facts (eliminate 2 conversions).

solution 2: Move all the special json encoding stuff to the new `cxx` bridge types. hackc's Facts types can now work with bincode, and are unconstrained by any json schema or `cxx` bridge constraints.  There is a single one-way conversion from hackc Facts to cxx bridge FileFacts that's about as efficient as such a thing can be; simply iterating & moving values from one container type to the other; without any cloning.

Conclusion: with this diff, step 1 & 2 become:
1. hackc::facts::Facts -> compiler_ffi::FileFacts
2. compiler_ffi::FileFacts -> json string

Reviewed By: nt591

Differential Revision: D51760152

fbshipit-source-id: 737a4736892ea01dae0fc84aefa4e004dffd2cfa