Fix some ProtocolReaderStructReadStateTest uses
Summary:
Previously, the reads:
- Weren't resilient w.r.t. advance failures caused by IOBuf boundaries.
- Didn't consistently pass state to protocol methods.

I guess arguably neither of these are "bugs" (these are all undocumented
internal APIs after all; maybe we can rely on default IOBufQueue sizes), but
both are easy to fix and the latter interferes with some subsequent performance
optimizations.

This is another in the slew of minor changes that will enable performance
optimization.

Reviewed By: iahs

Differential Revision: D46992082

fbshipit-source-id: bcc81a9c53b30d6edc7444afde47dff102d90934