Prefer direct interface implements to those inherited via extends
Summary:
With this diff, we start ancestor traversal with `implements` parents before `extends` parents.

Citing the long comment in this diff:
> A type may implement an interface or use a trait at multiple instantations, e.g. I<int> and I<string>. The [ancestor traversal] order determines which one will be used: we'll use the first one encountered during traversal.
Since this is a recursive function, ancestor traversal is in effect a depth-first traversal.

So by starting traversal with `implements` parent, we change which instantiation will win. The rest of the comment explains why we want to start with `implements` parents:
> For interfaces, the typechecking code forces the winner instantiation to be a subtype of each of the loser instantiations. This is necessary for soundness. Users can influence which instantiation wins by reordering direct parents.
It's important that we start traversal with interfaces here, otherwise interfaces which are ancestors of the base class would always win, with no easy way to override them.

Note that the part "the typechecking code forces the winner instantiation to be a subtype of each of the loser" will be implemented in the next diff.

That long comment also says:
> /!\ For soundness, the order here must be consistent with the traversal order for method folding!

Well starting with `implements` parents actually make the traversal order more consistent with the traversal order for member folding. See tests: one of the type hole tests now has errors.

Reviewed By: andrewjkennedy

Differential Revision: D35932600

fbshipit-source-id: 723b2cd670f8f234ffc5080238c671f0b065ddd8