templatize NamedEntity::get
Summary:
The allowCreate parameter is always constant at callsites; this gets
rid of a runtime branch in favor of specialized functions.

For correct case symbols, we need to reduce false positive logging
when someone dynamically looks up a missing symbol, with the side effect
of creating a named entity entry for the wrong-case, that later collides with
a correct-case symbol.

New getOrCreate() and getNoCreate() wrappers to make it obvious
when you are creating empty entries as a side effect. Followup diffs
will migrate dynamic symbol lookup call sites to getNoCreate().

Reviewed By: aorenste

Differential Revision: D51939928

fbshipit-source-id: 95595d1b5ec52e17fdfd1b38b9200ded9a034a2c