Experimental feature: FunctionRef
Summary:
Introduce an experimental feature flag `function_references` to support the typing of function pointer expressions using `FunctionRef<t>` instead of `t`, where `t` is a function type.

The idea is that whenever an API wants to make use of a function pointer in order to *update* it (typically via `fb_intercept2`), then it should take a FunctionRef, which correctly and soundly models a function *reference* because it is invariant in its type parameter. `FunctionRef` is defined simply using a newtype:
```
newtype FunctionRef<T> as T = T;
```
The bound of `T` ensures that any function reference can be used as an ordinary function value (e.g. for calling), but function values should *not* be used as proxy for a reference to the function, as this is unsound.

Note that lambdas do not have function reference types because they cannot be updated.

If the function is generic, we create an ordinary function value with some instantiation of the generic parameters. If in future we support first-class generic function types, then we could lift this restriction.

The feature has no runtime impact: `FunctionRef` is erased during emitting of bytecode.

Reviewed By: vassilmladenov

Differential Revision: D51848147

fbshipit-source-id: 3c525465eab29217212ac6a9f4b047f36d87179f