Make diagnostic formattable
Summary: Make `diagnostic` formattable by replacing the `ostream` inserter with a `formatter` specialization. This way it can be formatted with `fmt::format`, printed to a C or C++ stream or easily converted to a string. This also solves the synchronization problem when printing diagnostics from multiple threads as in `compile-thrift-files`.

Reviewed By: Mizuchi

Differential Revision: D51964523

fbshipit-source-id: c394f46821d744495434d04c2e0fb01622ff16cf