Move AnnotType::Object out of AnnotMetaType::Precise
Summary:
Make space for a new AnnotType that will represent any object type (aka
"object" typehint in systemlib, used for things like spl_object_hash()).

Since it makes sense for that one to be implemented using the
AnnotMetaType::Precise mechanism and name it AnnotType::Object, this diff
renames the AnnotType::Object to AnnotType::SubObject and creates a separate
AnnotMetaType::SubObject.

All locations where we check for AnnotType::Object and AnnotMetaType::Precise
were updated and their users were reviewed recursively.

It is a pretty boring diff, the most interesting change being in
TypeConstraint::underlyingDataType(), where the isPrecise() check no longer
covers sub-objects.

TypeConstraint::isAnyObject() was added for the future AnnotType::Object check
and it was used in a couple of places that would be otherwise easy to miss in
the next diff. It currently returns false.

If we later introduce a class<Foo> typehint, we should probably create an
analogous AnnotType::SubClass.

Reviewed By: mdko

Differential Revision: D51994060

fbshipit-source-id: 3853f787bfad7cf854112ec37f235e5b7cea6e48