store optionality of cpp.ref fields
Summary:
Currently in always-on reflection we determine optionality of a field using field_ref type. (i.e. `optional_field_ref`, `optional_boxed_field_ref`) For cpp.ref fields, we will not be able to determine the optionality on always-on reflection; therefore, we need to store this.

This only adds overhead for cpp.ref with optional fields, which is not expected to be expensive comparatively.

Reviewed By: iahs

Differential Revision: D51998780

fbshipit-source-id: 602f0aa9f7d178d64ee56d4e42c07d4eb98a2aac