Add support for "HH\object" type constraint
Summary:
Some builtins, such as spl_object_hash(), accept arbitrary objects.

Expose ability to express that with a "HH\object" typehint that gets mapped to
a newly introduced AnnotType::Object, which uses the AnnotMetaType::Precise.

The new typehint is not yet used anywhere except of the added unit test, as
"object" typehints are not automatically imported into the HH namespace.

Reviewed By: mdko

Differential Revision: D52018664

fbshipit-source-id: 4ce1a2be65967f223bf6a088c49bad56cc102894