Thread through CodeSource to parser
Summary:
When we start to ban utf8 in source code we'll need to distinguish between source code coming from a user file and source code in an eval so we can treat them differently for updating.

While here I also added variants for code coming from the debugger and code coming from systemlib - just in case we need that in the future as well.

Reviewed By: ricklavoie

Differential Revision: D51998618

fbshipit-source-id: ff250d753ef85b081cd29df41689e0b443b73cc9