Add logging for non-utf8 use in `eval`
Summary:
Add logging when eval compiles a non-utf8 source string.  The intention is to ship this to sandbox environments and turn on logging so we can see if/where www is using utf8 in dynamic eval strings.

The logging is controlled by `-vEval.EvalNonUtf8SampleRate`. The default is 0 which means no logging.

The results are stored in the `hhvm_isame_collisions` table with the event type `utf8`.

Reviewed By: ricklavoie

Differential Revision: D51998617

fbshipit-source-id: f790dd5fcc624c237a3054e3aae999ee78de04a8