Topologically sort children when traversing in process_roots in BSL (2/2)
Summary:
Topologically sort the children we call `aggregate_data` on.

This was not obvious to me last week. But topologically sorting the children is strictly better than just partitioning the children by "top". This is most easily seen by considering the case where our children are A and B, where A is a parent of B, and both are top nodes. Supposed B is before A in our list. We should just do 1 intersection with A, instead of intersecting with both A and B.

## Observations

When experimenting with greedy DFS BSL scheduling algo, I noticed computing `HH\IMemoizeParam` (where we promote any of its children to top that can be) was much longer than just computing `HH\IMemoizeParam` as the only top in the job.

One observation is that when we group related top inputs, the work can't be totally reused (ie computing a child and a parent is a decent amount more work than just computing a parent). In order to make the greedy DFS a win, we need to find a way to decrease this overhead.

Reviewed By: ricklavoie

Differential Revision: D51580465

fbshipit-source-id: 1a47f9bb15e852ed1a69f303be0bd31f3a32416d