Fix apc-locking.php to actually test locking again
Summary:
In D43916427 this test was updated to stop using the autoloader but no longer tests the reentrancy bug in APC. While we no longer have user land autoloading it is still possible to reenter during apc unserialization via `__wakeup` methods and error handling (the unserializer can raise notices and warnings).

Update the test to use a `__wakeup` method to produce the same reentrant behavior as before.

Reviewed By: ricklavoie

Differential Revision: D52146241

fbshipit-source-id: 55d3bc7c2b24558452919607abe2935885a25862