Check constructor bodies under dynamic assumptions
Summary: In a class marked `<<__SupportDynamicType>>` we check methods under both static and dynamic assumptions. But we check constructors only under static assumptions. This is sound if the class cannot be *instantiated* dynamically, but in the presence of `new $x` where `$x` has type `dynamic`, to be sound we must check the constructor under dynamic assumptions.

Reviewed By: vassilmladenov

Differential Revision: D52052843

fbshipit-source-id: f5e0d10adb0599d048c1bc6cf09b45c06fb42e66