initialize CAS client on server load
Summary:
We've found in practice initialization of the CAS client to take on average around 8 seconds. Let's move this earlier in the server init process so we can do it in parallel with other work like saved state downloading. This uses a similar paradigm as hh_distc where we spawn a thread and then us a mspc channel to communicate b/w hh_server and the spawned thread with the actual CAS client.

For reviewers, I'd like to get this out before the code freeze to gather telemetry so if there's no correctness/performance concerns, I'd prefer if we could land this earlier rather than later and stack on clean up diffs.

Reviewed By: edwinsmith

Differential Revision: D52046312

fbshipit-source-id: 850a9f3cbde964ed5959aec6a5a8575bd16b690d