Do not access the OCaml runtime in hh_should_exit
Summary:
Accessing Caml_state requires the domain lock. The domain lock is held by one
thread at a time. The local root macros (CAMLparamX, CAMLlocalX, CAMLreturn)
require the domain lock to be held. If it isn't, we see the fatal error "no
domain lock held"

The Glean code in Hack calls hh_should_exit from threads which do not hold the
domain lock. This is a fault and crashes under OCaml 5.1. We could fix this by
always running hh_should_exit from the "OCaml thread" in this code, like we
already do for the raising version of that function.

However, we can avoid requiring the domain lock entirely, because we don't
actually need access to the OCaml runtime in this function, as it does not
allocate or raise.

To fix, I've marked this function as "noalloc" which also makes calls from OCaml
cheaper.

Differential Revision: D52221930

fbshipit-source-id: bd67a592f8d7b19fdc07882664660ca589e6fd54