Work around lack of major heap chunks GC stat in OCaml 5
Summary:
In OCaml 5, `Gc.stat ()` returns a record with a `heap_chunks` field, but the
value is always 0 because this statistic is not tracked in OCaml 5.

See https://v2.ocaml.org/api/Gc.html

This change removes an assertion which can't possibly pass under OCaml 5, but I
made no effort to update the rest of the code. I'm not sure how this code relies
on the heap chunks number, but my hope is that any further changes can be made
later after OCaml 5 lands.

Reviewed By: edwinsmith

Differential Revision: D52222403

fbshipit-source-id: 0168ce53a9865ede74af4adf2d6af2260fc22924