Expression IDs have class scope instead of method scope
Summary:
The previous diff D50271708 was a refactoring of how we handle expression IDs.

It made is so that:
* The expression ID provider is reset for each method. Before that diff, it was reset for each class with the display map reset for each method.
  * This was to have the same behaviour/test outputs as before, but I don't like it because
    * when doing TAST checks, we don't reset the provider for each method. So we might have <expr#1> in multiple methods (but not referring to the same expression), but the display ID map will be the same across methods, which might give odd numbering in error messages.

So in this diff, I make expression IDs have class scope. You can see the effect on error messages in test outputs. For example, one test has it's first error message start at <expr#6> instead of <expr#1>. As we explained in the previous diff, this is due to the display map being used for other things than error messages, including string mingling for `Tgeneric`.

Reviewed By: andrewjkennedy

Differential Revision: D52239835

fbshipit-source-id: 48ce4058f83c9f190d72d85a58fdbf6242f51969