Better names for Counters telemetry
Summary:
We have a variety of forms of telemetry. One is Counters.ml, for some ultra-lightweight counters+timings of certain frequently-called functions - particular Decl_provider.get_*.

I've always found the labels used in this telemetry to be completely baffling. (and it was I who wrote them in the first place!) This diff renames them, and rewrites the comments.

Prior to this diff, some of the functions measured wall-time and some measured cpu-time. That split was born of a hope that maybe cpu-time would be stable even across different processors, since we observed that wall-time varied a lot. Unfortunately cpu-time also varied just as much. Also, having an IO operation Disk_cat measured in wall-time (since cpu-time is meaningless for IO), while functions like Direct_decl_parse which call into Disk_cat were measured in cpu-time, meant that things were hard to add up. Therefore, this diff simplifies things by switching everything over to wall-time.

Here are the renamed categories and comments in this diff.
* `Decling -> Direct_decl_parse`
   * This is counted every use of [Direct_decl_utils.direct_decl_parse{_and_cache}]. It often includes some [Disk_cat] time.
* `Get_decl -> Decl_provider_get`
   * This measures every [Decl_provider.get_*] function. In cases where it's not already in the cache, this will also include calls to [Direct_decl_parse] hence also [Disk_cat], and also time to do decl-folding.
* `Get_ast -> Ast_provider_get`
   * This measures every [Ast_provider.get_ast_with_error] call, for fetching the full ASTs of the files we're typechecking. It is inclusive of [Disk_cat] time. Note that this is *not* used for direct-decl-parsing. *)
* `Typecheck -> Typing_toplevel`
   * This measures every [Typing_top_level] call. It will often include [Decl_provider_get] and hence [Direct_decl_parse] and [Disk_cat] for all the decls it needs.

Differential Revision: D52182574

fbshipit-source-id: 9922d95059f9dfa99e1733e256124d40919b82ee