Greedy DFS BSL algorithm
Summary:
The idea here is to avoid duplicate work by grouping related classes.
## The approach
We first separate out the full class list into `roots`, `rootLeafs`, `split`, and `child`.
- `roots` are used to know where to start traversing from
- `rootLeaf` are the disjoint classes
- We run `dfs_bucketize` on the roots, which is just a DFS. When we find a node with less than X transitive dependencies, we greedily add it and all its children to our bucket. If a bucket has more than `kMaxBucketSize`, consistently bucketize into buckets that are roughly of cost `kMaxBucketSize`.
- We consistently hash the `rootLeaf`
- We run `assign_hierarchical_work` on both sets of buckets, taking care to add atleast the immediate deps, (or more if not yet processed) to these buckets.
## Follow-ups / Observations

To further improve, we could tune the constants I set here.
We could also create bigger buckets to consistently bucketize if we find this is not consistent enough.

## Questions
Q: If class A extends B. And they both need to be processed. Does it ever make sense to ever process them in separate batches???
A: Previously we were recalculating everything! Now we should just be calculating B once, and then using that for A, so now it never really makes sense. Of course, there is a bit of extra time to iterate over child lists, even if we have already computed everything, but added time _should_ be negligible.

Q: Whats the difference between this diff and the approach of just promoting more than just leafs?
A: For the algorithm that just promotes thats not a leaf, we would choose frontier set such that the transitive dependency count is less than some threshold, and consistently assign those to buckets. Then, we'd go to add all the transitive dependencies. This would result in likely more uneven distribution since we didn't take into account the actual transitive dependency sizes while doing the consistent bucketing.

Q: Whats the tradeoff as to whether this diff is a win?
A:

Good:
- Decreasing total work isn't really one of our explicit goals, but could hypothetically help because since we have limited threads, other jobs would get to run earlier.
- fewer jobs means less total job overhead
- more overlapping dependencies means less time deserializing inputs
- makes the buckets more even.

Bad:
- Makes the buckets less consistent
- more output to serialize and store from each job since there is a higher ratio of top to deps

Reviewed By: ricklavoie

Differential Revision: D51927698

fbshipit-source-id: 1825e63d55ca114ac36fd83267b12279a1476a1e