DDP: Always keep heredocs/nowdocs/parenthesized expressions
Summary:
I'm undoing D32299873. That diff aimed at discarding heredocs/nowdocs from attributes whenever we kept them, but had the unintended effect of removing all heredocs/nowdocs whenever we kept attributes.

It was done so to preserve the old behaviour when migrating the fact parser, but right now this diff does not seem to break anything.

Differential Revision: D52229862

fbshipit-source-id: b2ed08289343290979e74a3b3d0172d8c277dc03