Remove expression ID simplification for errors
Summary:
As explained in previous diff D50271708:

In Typing_reason, we maintain a map `expr_display_id_map` from expression IDs to integers that will be used to display expression IDs in errors (so instead of displaying some seemingly random <expr#1234>, we display some nice small and consecutive numbers.

However, **the way this map is used is extremely dodgy**. When comparing TAST hashes and errors (e.g. for mass replays), I sometimes see random discrepancies in expression IDs. When I remove the display map entirely, I get rid of that non-determinism.

Why it's dodgy:
* because it's not just used for errors, it's used in particular by DependentKind.to_string, which flows into all sorts of things, including `Tgeneric`. That leads to horrendous logic like this one in Typing_escape: https://fburl.com/code/08z3wyk1. Quentin was smart enough to spot the complexity, but surely the subtlety was missed in other places.
* The observed non-determinism is a smell.
* It's hard to make sure the display map is reset properly. Before D50271708, the map would be reset for each method, but then stay as it was for the last method for TAST checks, and possibly other paths.
* Using the display map for DependentKind.to_string defeats the goal of cleaner error messages. It's frequent that errors don't start with <expr#1>.
  * I've tried to create another flavour of DependentKind.to_string which does not use the map, but that leads to inconsistent error messages (e.g. "but got <expr#23>::T ... where <expr#1> refers to this expression"). (My attempt: D52253238).

**I think we should remove the map**, trading simplicity and likely correctness for a bit more quirkiness of error messages, which I think is a fine trade.

This is what this diff does.

Reviewed By: andrewjkennedy

Differential Revision: D52253783

fbshipit-source-id: 6546d8b7682f658b458cff5200e9117545baf9d4