Use new type Tvid.t for type variable IDs
Summary:
So far, type variables IDs are still ints in many parts of the code. Now that we have `Ident_provider`, use it to create a new abstract type Tvid.t, still guaranteeing determinism of the IDs but also abstraction and separation from other kinds of IDs.

Just like expression IDs have a provider in the env, Tvid also does (`env.tvar_id_provider`)

The diff is quite mechanical, I'm not sure there are any tricky bits.
It's a lot of these replacements:
* `Ident.t`/`int`/`Typing_inference_env.identifier` -> `Tvid.t`
* `IMap`/`ISet` -> `Tvid.Map`/`Tvid.Set`

Differential Revision: D52237462

fbshipit-source-id: 179338e220f39292355eee2eb9fdd8a89e6f55d3