Decl_provider.prefetch - prefetch ancestors during declare_folded_class
Summary:
This diff changes the implementation of `Decl_provider.get_class` for the Local_memory backend, which is used in the IDE. New behavior:

> If we call Decl_provider.get_class, and the folded decl isn't already present, and the JK `lsp_prefetch_decls` is turned on, then do a multicore "prefetch" of all the shallow decls that Decl_provider.get_class will need, before proceeding.

(the following abandoned diff D52240067 attempted to be more ambitious, to prefetch as many decls for the entire file as it can guess rather than just for the current declare_folded_class; but it didn't bring any additional perf benefit so I abandoned it.)

Which shallow decls are truly needed, and which ones does the "prefetch" in this diff provide? More specifically, we want to prefetch the decls which will be needed for typical *use* of the decl, e.g. to do subtyping, looking up instance methods, look up static methods; for classlikes, this winds up the same as folding the decl.
```
FOLD                  TRUTH    PREFETCH   UNDER   OVER
\:b                      228        229       0      1
\CalendarClient          106        106       0      0
\ChessRound **             4         60       0     56
\EntChessLeague          602        603       0      1
\EntGroup              5,995      5,995       1      1
\EntBusinessUser      10,617     10,623       0      6
\EntBusinessUserSpec   1,003      1,003       0      0
\EntBusinessUserAccessors
                         589        585       4      0
\EntBusinessUserCreationMutationBuilder
                      10,017     10,018       0      1
```
TRUTH column says how many decls (shallow decls, typedefs, funs, consts, modules) are truly needed by `Decl_provider.get_{class,typedef}` to fetch the folded decl or the typedef.

PREFETCH column shows how many decls were fetched by the code in this diff

UNDER column shows how many decls this diff failed to fetch, that were nevertheless truly needed. False negatives.

OVER column shows how many decls this diff fetched that were not truly needed. False positives.

`\ChessRound` is a typedef. The "TRUTH" column isn't much use for that, because `Decl_provider.get_typedef` merely fetches the shallow decl. This doesn't help for typical use (e.g. subtyping, or looking up instance methods). This diff aims to fetch the decls that will be needed for typical use.

## How does prefetching work?

To prefetch for a folding a decl `class D extends A,B,C {}` that's not already in folded cache, we start by looking at the shallow decl for D (either already in cache, or direct-decl-parsing off disk). This tells us that we'll also need folded decls for A,B,C, so we kick off these prefetches in parallel. Wide ancestor hierarchies are very amenable to parallelisation!

Observe that the algorithm repeatedly switches between cache-side work and direct-decl-parsing work:
1. Walk as far as we can in existing shallow+folded caches to find what shallow decls will be needed next
2. Resolve those to filenames
3. Kick off concurrent async direct-decl-parsing for those files
4. As soon as we got a direct-decl-parse from an earlier concurrent decl-fetch, store it in cache, and repeat from step 1

I was working for IDE benefits, where the caches are all stored in ocaml memory as HashTbl. I therefore figured it made most sense to implement step 1 in ocaml. If this same technique were to be used for the rust backend for hh_server then the equation might be different, but I never got to the bottom of what is the storage for those caches.

Step 2 could have been done either in Rust or Ocaml; it might have been nice to do naming-sqlite lookups concurrently on the rust side, rather than sequentially on the ocaml side, but that would be a bunch more engineering.

## Challenges

**Pessimisation**. When you fold a decl, the first step is to look up all the ancestors, and the second step is to dig through the return types of every method in the decl itself (not its ancestors) and judge whether they are enforced or not per the sound dynamic rules. Digging through the return type involves transitively resolving typedefs, and fetching further decls. Pessimisation contributes towards drastically higher numbers for "TRUE required decls". The precise algorithm for pessimisation is complicated. I really wanted a simple approximation. I iterated several times until I got something that was simple and had close enough agreement to the needs of pessimisation.

**Enums**. The folding rules for enums and enum classes are a bit complicated, in terms of which decls they need. For instance with `enum E {}` the shallow class decl for E indicates that it extends `\HH\BuiltinEnum`. To fold E, do we need the shallow or folded decl for this? If we subsequently want to look up an instance member on E, or do subtyping on it, how about then? What about `enum E : F as G`? And enum classes? I iterated several times until I got something that was simple and had close agreement to the needs of enum folding.

Differential Revision: D52018688

fbshipit-source-id: 5607021c06798769a04dc292860721654e9b0704